poisonwiki/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
black==24.4.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2024.2
cryptography==42.0.8
cssbeautifier==1.15.1
decorator==5.1.1
distro==1.9.0
dj-database-url==2.2.0
Django==5.0.6
django-active-link==0.2.1
django-admin-interface==0.28.8
django-allauth==0.63.3
django-authtools==2.0.1
django-cleanup==8.1.0
django-colorfield==0.11.0
django-cors-headers==4.4.0
django-crispy-forms==2.2
django-currentuser==0.7.0
django-environ==0.11.2
django-extensions==3.2.3
django-extra-views==0.14.0
django-filter==24.2
django-htmx==1.18.0
django-localflavor==4.0
django-minify-html==1.8.0
django-pwa==2.0.0
django-seed==0.3.1
django-tables2==2.7.0
django-view-breadcrumbs==2.5.1
djangorestframework==3.15.2
djlint==1.34.1
dotenv-stripout==0.1.3
EditorConfig==0.12.4
executing==2.0.1
Faker==26.0.0
gunicorn==23.0.0
h11==0.14.0
html-tag-names==0.1.2
html-void-elements==0.1.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipython==8.26.0
isort==5.13.2
jedi==0.19.1
jsbeautifier==1.15.1
json5==0.9.25
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
minify_html==0.15.0
mypy-extensions==1.0.0
openai==1.35.10
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyOpenSSL==24.1.0
pyreadline==2.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
python-stdnum==1.20
PyYAML==6.0.1
regex==2023.12.25
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
stack-data==0.6.3
text-unidecode==1.3
toposort==1.10
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
whitenoise==6.7.0
django-import-export==4.1.1
